Rename `index` entrypoint to `sprintf`
